Github Readme Makeover Step 5
Changed <a> tags for social buttons to markdown format to get rid of weird blue lines behind & next to buttons that was being caused by the <a> tags. Added a button for my portfolio site that users the dragon icon from Dungeons and Dragons. 

All icons for shield.io badges are sourced from Simple Icons.